add header for makefile

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: imqandyl <imqandyl@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/03 11:34:34 by imqandyl          #+#    #+#             */
-/*   Updated: 2024/07/03 11:34:34 by imqandyl         ###   ########.fr       */
+/*   Updated: 2024/07/03 23:19:40 by imqandyl         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
